Support naive datetimes from server. Fixes #109
Adds a new `default_timezone` attribute to `Account` which is used to localize naive datetimes that are returned from the server. Defaults to the timezone of the host. Make an exception for attachments, where the last_modified_time is apparently always in UTC.
This required changing the signature of the .from_xml() method which is used in many places.