Simplify handling of timezone fields
Make these fields real item fields, but name them with an underscore to make it clear that these fields are not meant for general consumption. They are only used to pass timezone data related to start and end fields on CalendarItem. The change allows us to remove ugly code in the UpdateItem service.

Add an optional 'deprecated_from' attribute to the Field class to handle the deprecated MeetingTimeZone field. It's not deprecated in the official docs, but some Exchange versions > 2007 issue an error if you try to use it.

While here, make datetime tests more robust by sidestepping a Microsoft timezone bug on the Europe/Copenhagen timezone.