First attempt at single query on multiple folders
Create a FolderCollection class to collect groups on folders, and implement the query API on top of that. Keep the query API on Folder class but make it a thin wrapper of Folder Collection methods.

Currently fails on multiple folders because Service._get_page() expects only one result element container but gets multiple. We need to implement paging for each result element individually.

Refs #224