Fix fields for DeleteFolderChange
As a subclass of FolderChange, DeleteFolderChange was previously looking
for a 'Folder' field when parsing the xml response. However, this field
does not exist for folder deletions. 'FolderId' is the only field
returned. This diff updates DeleteFolderChange to parse the "FolderId"
field into an `item_id` attribute. It takes the form of a raw tuple with
the id and changekey.

See https://docs.microsoft.com/en-us/previous-versions/office/developer/exchange-server-2010/aa565228%28v%3dexchg.140%29