Work around EWS bug updating datetimes. Fixes #310
If we create a calendar item with datetimes in a local timezone and then
request the item again, we will get the datetimes as UTC. This is expected
as we rely on the user to convert to the correct timezone when presenting
datetimes to the user.

But if we save() the item we just fetched, we now send the datetimes as UTC,
and EWS may shift the datetimes (by an hour in my tests, but possibly more).
This is argually a bug in EWS, but we'll have to work around it. Do so by
adding localized datetimes in the XML. This maintains backwards compatibility
by keeping datetimes on the item instance as UTC.