Support multiple folder hierarchies
It turns out archive and public folders reside in entirely separate folder hierarchies.

Change Folder class to take a root folder instance instead of an account instance and
assuming the folder belongs to account.root. Create a new RootOfHierarchy folder that
holds a reference to the account and knows how to traverse and cache subfolders.

Split the WELLKNOWN_FOLDERS collection and move to the RootOfHierarchy folder subclasses
where they belong.

Refs #396