Timezone map fixes (#521)
* Remove incorrect hard-coding of "GMT" for "GMT Standard Time"

The hard-coding had the effect of picking the pytz timezone GMT for
events in the MS timezone "GMT Standard Time", which is wrong - we
need a pytz timezone where there is UTC+0 in winter and UTC+1 in
summer. Removing this mapping means that a more correct tz will be
selected.

This is a partial fix for #519

Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>

* Store the territory associated with each timezone mapping

For each CLDR timezone, there is a territory; one entry corresponding
to each Microsoft timezone is the default mapping, which is territory
'001' in addition to its usual territory. So store the territory in
the CLDR_TO_MS_TIMEZONE_MAP (keeping the 001 territory by
preference).

Then, when generating the reverse map MS_TIMEZONE_TO_PYTZ_MAP, take
the mapping with territory 001.

This is a partial fix for #519

Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>

* Reflect changed structue of PYTZ_TO_MS_MAP

The values in PYTZ_TO_MS_MAP are now tuples of (timezone name,
territory); so the lookup code needs adjusting to take the first
element of the tuple.

Also update tests to reflect the revised structure.

This is the final part of the fix for #519

Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>