Improve the way we search for folders
Add a new FolderQuerySet that uses the same principles as the QuerySet class. Use this new
class everywhere we want to search, filter or traverse the folder structure from a given place
in the tree. This hides the low-level find_folders() and get_folders() methods and replaces
them with easier-to-understande .all(), .get() and .filter() methods for folder traversal.

While here, add a new FolderCollection.resolve() method that calls GetFolder on a set of folder
IDs to get the full set of field values on these folders (e.g. distinguished folders).