Be less smart when clearing XML elements
If we happen to parse the child elements in a different order than they
arrive in the response XML, then elem.getparent()[0] points to an element
that we have not yet handled. That element would then disappear. Instead,
only remove from the parent the one element we have just parsed. Don't make
other assumptions.

Fixes #585