Implement support for OccurrenceItemId
This identifier field allows us to indicate a specific occurrence within
a recurring series so that we can make individual changes to it. You can
also update occurrences with their direct item id, but unless you're
storing every single occurrence on your end, you're not going to have
the item id for every occurrence. The OccurrenceItemId identifies an
occurrence via the master event's item id and the instance number of the
occurrence within the series.