Support OAuth authorization code grant type (#698) (#699)
* Support OAuth authorization code grant type (#698)

Previous work to use OAuth added support for the client credentials
grant type, which can be useful for an internal application intended
only to be authorized to access one organization's data. However, it
doesn't work well for multitenant applications since there's no way to
guarantee that the redirect from authorization wasn't tampered with,
which could enable an attack involving changing the tenant ID before
completing the redirect.

That's precisely the problem the authorization code grant solves, so
this expands OAuth support to cover that grant type. This includes a
provision for a somewhat specific case where the calling application
isn't privy to the OAuth application ID and secret and instead relies
on another service to issue and refresh access tokens. In such cases,
applications can subclass OAuth2AuthorizationCodeCredentials and
implement refresh() to provide their own refresh logic.

* Fix Python 2 incompatibility, parameter mismatch, Bandit false positive, lock reentry (#698)