Add a better autodiscover implementation
Reimplement the autodiscover protocol in a way that is closer to the description
of the protocol. It is also better documented, has fewer nesting and is thus easier
to reason about and debug.

Keep the old implementation around for now. It has some hacks that were necessary
to autodiscover misconfigured sites.

Add testing of the new implementation.