Update post_ratelimited to better handle unicode data
We have this pretty hairy UnicodeDecodeError bug going on:
https://app.clubhouse.io/nylas/story/18092/event-creation-fails-in-syncback-with-unicodedecodeerror

It looks like we have a string with some utf-8 characters in it, but the
string isn't specified as unicode.

My first attempt to fix this was just to call `.decode('utf-8')` on the
data before we passed it down into requests, but that came back to us
with a schema validation error from the provider.

After lots of digging, turns out the Content-Length header was incorrect
:face-palm: Requests doesn't have good enough unicode support to realize
it should calculate the byte length instead of the number of characters.
So, I've rejiggered our request building so that we can calculate the
content length ourselves.