Try different encoding approach in post_ratelimited
We were getting some SSL bad write retry errors, and mostly the internet
just seems to think that passing in unicode is a bad idea. This tries to
encode the data instead. Try utf-8 first, and if that fails we see if
chardet can get the encoding for us. Also had to encode the
authorization header because it was causing the data body to be coerced
(unsuccessfully) to unicode.